represent each digit as a bitset
a is the 1st bit, g is the 7th bit

the rest can be done by counting the number of bits in the resultant
numbers. the tricky part of this is when you have to remove bits from a
number, B, from another number, A. to do this, we can need the following
truth table for each bit:

  | 0 1  A
--+------
0 | 0 1
1 | 0 0
B

which works out to flipping all the bits in B (1s complement), then performing an AND
operation. or in ruby land... `A & ~B`